chore: Update HttpDownload class to support direct download from a URL
This commit updates the HttpDownload class to add support for triggering a direct download from a URL. The `triggerDirect` method has been added to the `download` object in the `useFetch` composable. This allows users to initiate a download directly from a URL without having to make a request and receive a response.

The `triggerDirect` method takes a URL and an optional filename as parameters. It internally calls the static `direct` method of the `HttpDownload` class to handle the download process. If no URL is provided, a warning message is logged to the console.

This enhancement improves the usability of the `useFetch` composable by providing a convenient way to trigger downloads from URLs. It also enhances the flexibility of the `HttpDownload` class by allowing direct downloads without the need for a response object.

Closes #123